about delete

+++ src/Components/Admin/AdminAbout/AdminAbout.jsx
@@ -1,3 +1,4 @@
+import { useQuery } from "@tanstack/react-query";
 import React, { useEffect } from "react";
 import { useState } from "react";
 import { useForm } from "react-hook-form";
@@ -6,6 +7,12 @@ import About from "../../About/About";
 const AdminAbout = () => {
   const [aboutDesc, setAboutDesc] = useState([]);
   const { register, handleSubmit, reset } = useForm();
+
+  useEffect(() => {
+    fetch("http://localhost:5000/admin/about")
+      .then((res) => res.json())
+      .then((data) => setAboutDesc(data));
+  }, [aboutDesc]);
 
   const handleAddDesc = (data) => {
     const desc = data.desc;
@@ -31,23 +38,25 @@ const AdminAbout = () => {
       .catch((err) => console.error(err));
   };
 
-  const handleDeleteAbout = (about) => {
-    fetch(`http://localhost:5000/about/${about._id}`, {
-      method: "DELETE",
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        if (data.deletedCount > 0) {
-          alert.success(`${about?.name} deleted successfully`);
-        }
-      });
-  };
+  const handleRemove = (id) => {
+    const proceed = window.confirm(
+        "Are you sure, you want to remove this order?"
+    );
+    console.log(proceed)
+    if (proceed) {
+        fetch(`http://localhost:5000/aboutDesc/${id}`, {
+        method: "DELETE",
+        })
+        .then((res) => res.json())
+        .then((data) => {
+            if (data.deletedCount > 0) {
+            alert("Removed Order Successfully");
+            }
+        });
+    }
+};
 
-  useEffect(() => {
-    fetch("http://localhost:5000/admin/about")
-      .then((res) => res.json())
-      .then((data) => setAboutDesc(data));
-  }, [aboutDesc]);
+console.log(aboutDesc)
 
   return (
     <section className="w-full ml-16 ">
@@ -156,70 +165,64 @@ const AdminAbout = () => {
         </div> 
         </div>*/}
  
-        <div class="overflow-hidden rounded-lg border border-gray-200 shadow-md m-5">
-          <table class="w-full border-collapse bg-white text-left text-sm text-gray-500">
-            <thead class="bg-gray-50">
+        <div className="overflow-hidden rounded-lg border border-gray-200 shadow-md m-5">
+          <table className="w-full border-collapse bg-white text-left text-sm text-gray-500">
+            <thead className="bg-gray-50">
             </thead>
-            <tbody class="divide-y divide-gray-100 border-t border-gray-100">
-              <tr class="hover:bg-gray-50">
-                <th class="flex gap-3 px-6 py-4 font-normal text-gray-900">
-           
-                  <div class="text-sm">
-                    <div class="font-medium text-gray-700 ">
-                      Lorem Ipsum is simply dummy text of the printing and
-                      typesetting industry. Lorem Ipsum has been the industry's
-                      standard dummy text ever since the 1500s, when an unknown
-                      printer took a galley of type and scrambled it to make a
-                      type specimen book. It has survived not only five
-                      centuries, but also the leap into electronic typesetting,
-                      remaining essentially unchanged. It was popularised in the
-                      1960s with the release of Letraset sheets containing Lorem
-                      Ipsum passages, and more recently with desktop publishing
-                      software like Aldus PageMaker including versions of Lorem
-                      Ipsum.{" "}
-                    </div>
+            <tbody className="divide-y divide-gray-100 border-t border-gray-100">
+             {
+              aboutDesc?.map(desc =>  <tr className="hover:bg-gray-50">
+              <th className="flex gap-3 px-6 py-4 font-normal text-gray-900">
+         
+                <div className="text-sm">
+                  <div className="font-medium text-gray-700 ">
+                  {desc?.desc}
                   </div>
-                </th>
+                </div>
+              </th>
 
-                <td class="px-6 py-4">
-                  <div class="flex justify-end gap-4">
-                    <a x-data="{ tooltip: 'Delete' }" href="#">
-                      <svg
-                        xmlns="http://www.w3.org/2000/svg"
-                        fill="none"
-                        viewBox="0 0 24 24"
-                        stroke-width="1.5"
-                        stroke="currentColor"
-                        class="h-6 w-6"
-                        x-tooltip="tooltip"
-                      >
-                        <path
-                          stroke-linecap="round"
-                          stroke-linejoin="round"
-                          d="M14.74 9l-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 01-2.244 2.077H8.084a2.25 2.25 0 01-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 00-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 013.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 00-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 00-7.5 0"
-                        />
-                      </svg>
-                    </a>
-                    <a x-data="{ tooltip: 'Edite' }" href="#">
-                      <svg
-                        xmlns="http://www.w3.org/2000/svg"
-                        fill="none"
-                        viewBox="0 0 24 24"
-                        stroke-width="1.5"
-                        stroke="currentColor"
-                        class="h-6 w-6"
-                        x-tooltip="tooltip"
-                      >
-                        <path
-                          stroke-linecap="round"
-                          stroke-linejoin="round"
-                          d="M16.862 4.487l1.687-1.688a1.875 1.875 0 112.652 2.652L6.832 19.82a4.5 4.5 0 01-1.897 1.13l-2.685.8.8-2.685a4.5 4.5 0 011.13-1.897L16.863 4.487zm0 0L19.5 7.125"
-                        />
-                      </svg>
-                    </a>
-                  </div>
-                </td>
-              </tr>
+              <td className="px-6 py-4">
+                <div className="flex justify-end gap-4">
+                  <button x-data="{ tooltip: 'Delete' }" onClick={() => handleRemove(desc._id)}>
+                    <svg
+                      xmlns="http://www.w3.org/2000/svg"
+                      fill="none"
+                      viewBox="0 0 24 24"
+                      stroke-width="1.5"
+                      stroke="currentColor"
+                      className="h-6 w-6"
+                      x-tooltip="tooltip"
+                    >
+                      <path
+                        stroke-linecap="round"
+                        stroke-linejoin="round"
+                        d="M14.74 9l-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 01-2.244 2.077H8.084a2.25 2.25 0 01-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 00-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 013.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 00-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 00-7.5 0"
+                      />
+                    </svg>
+                  </button>
+                  {/* <button onClick={() => handleRemove
+                  (desc._id)}>X</button> */}
+                  <a x-data="{ tooltip: 'Edite' }" href="/">
+                    <svg
+                      xmlns="http://www.w3.org/2000/svg"
+                      fill="none"
+                      viewBox="0 0 24 24"
+                      stroke-width="1.5"
+                      stroke="currentColor"
+                      className="h-6 w-6"
+                      x-tooltip="tooltip"
+                    >
+                      <path
+                        stroke-linecap="round"
+                        stroke-linejoin="round"
+                        d="M16.862 4.487l1.687-1.688a1.875 1.875 0 112.652 2.652L6.832 19.82a4.5 4.5 0 01-1.897 1.13l-2.685.8.8-2.685a4.5 4.5 0 011.13-1.897L16.863 4.487zm0 0L19.5 7.125"
+                      />
+                    </svg>
+                  </a>
+                </div>
+              </td>
+            </tr>)
+             }
             </tbody>
           </table>
         </div>
